requirements.txt
numpy==1.24.3
Pillow==11.0.0
streamlit==1.40.1
streamlit==1.32.0
streamlit==1.40.2
streamlit_option_menu==0.4.0
tensorflow==2.13.0
tensorflow==2.10.0
tensorflow==2.18.0
tensorflow_intel==2.13.0
tensorflow_intel==2.18.0
